requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0
Flask-Migrate==4.0.5
SQLAlchemy==2.0.31
python-dotenv==1.0.1
Werkzeug==3.0.1
reportlab==4.2.0
psycopg2-binary==2.9.9
pandas==2.2.2
openpyxl==3.1.2
Pillow==10.3.0
